+++ HimalayaMain/Master/Components/Scheduler/Commands/Include/CmdALFilling.h
@@ -138,7 +138,7 @@ public:
      *  \return from GetStrResult
      */
     /****************************************************************************/
-    QString GetStrResult()const{ return QString("%1").arg(m_result);}
+    QString GetStrResult()const{ return QString("%1").arg((qint32)m_result);}
 
 private:
     CmdALFilling();                                         ///< Not implemented.
