+++ HimalayaMain/Master/Components/Scheduler/Commands/Include/CmdALGetRecentTemperature.h
@@ -124,7 +124,7 @@ public:
 	 *  \return from GetParameters
 	 */
 	/****************************************************************************/
-	QString GetParameters()const{ return GetName() + QString("(%1,%2)").arg(m_Type).arg(m_Index);}
+    QString GetParameters()const{ return GetName() + QString("(%1,%2)").arg((qint32)m_Type).arg(m_Index);}
 	/****************************************************************************/
 	/*!
 	 *  \brief  Definition/Declaration of function GetStrResult
@@ -132,7 +132,7 @@ public:
 	 *  \return from GetStrResult
 	 */
 	/****************************************************************************/
-	QString GetStrResult()const{ return QString("%1").arg(m_result);}
+    QString GetStrResult()const{ return QString("%1").arg((qint32)m_result);}
 
 private:
     CmdALGetRecentTemperature();                                                    ///< Not implemented.
