+++ HimalayaMain/Master/Components/Scheduler/Commands/Include/CmdALStartTemperatureControlWithPID.h
@@ -214,7 +214,7 @@ public:
 	 *  \return from GetParameters
 	 */
 	/****************************************************************************/
-	QString GetParameters()const{ return GetName() + QString("(%1,%2,%3,%4,%5,%6,%7)").arg(m_Type).arg(m_NominalTemperature).arg(m_SlopeTempChange).arg(m_MaxTemperature).arg(m_ControllerGain).arg(m_ResetTime).arg(m_DerivativeTime);}
+    QString GetParameters()const{ return GetName() + QString("(%1,%2,%3,%4,%5,%6,%7)").arg((qint32)m_Type).arg(m_NominalTemperature).arg(m_SlopeTempChange).arg(m_MaxTemperature).arg(m_ControllerGain).arg(m_ResetTime).arg(m_DerivativeTime);}
 	/****************************************************************************/
 	/*!
 	 *  \brief  Definition/Declaration of function GetStrResult
@@ -222,7 +222,7 @@ public:
 	 *  \return from GetStrResult
 	 */
 	/****************************************************************************/
-	QString GetStrResult()const{ return QString("%1").arg(m_result);}
+    QString GetStrResult()const{ return QString("%1").arg((qint32)m_result);}
 
 private:
     CmdALStartTemperatureControlWithPID();                                                    ///< Not implemented.
