+++ HimalayaMain/Master/Components/Scheduler/Commands/Include/CmdCloseDiagnosticService.h
@@ -96,7 +96,7 @@ public:
 	 *  \return from GetStrResult
 	 */
 	/****************************************************************************/
-	QString GetStrResult()const{ return QString("%1").arg(m_result);}
+    QString GetStrResult()const{ return QString("%1").arg((qint32)m_result);}
 
 private:
     CmdCloseDiagnosticService();                                                    ///< Not implemented.
