+++ HimalayaMain/Master/Components/Scheduler/Commands/Include/CmdOvenGetTemperatureControlState.h
@@ -106,7 +106,7 @@ public:
 	 *  \return from GetParameters
 	 */
 	/****************************************************************************/
-	QString GetParameters()const{ return GetName() + QString("(%1)").arg(m_Type);}
+    QString GetParameters()const{ return GetName() + QString("(%1)").arg((qint32)m_Type);}
 	/****************************************************************************/
 	/*!
 	 *  \brief  Definition/Declaration of function GetStrResult
@@ -114,7 +114,7 @@ public:
 	 *  \return from GetStrResult
 	 */
 	/****************************************************************************/
-	QString GetStrResult()const{ return QString("%1").arg(m_result);}
+    QString GetStrResult()const{ return QString("%1").arg((qint32)m_result);}
 
 private:
     CmdOvenGetTemperatureControlState();                                                    ///< Not implemented.
